Revert "Add breadcrumbs where needed"
