+++ app/presenters/detailed_guide_presenter.rb
@@ -1,14 +1,28 @@
 class DetailedGuidePresenter < ContentItemPresenter
-  include Breadcrumbs
+  include Political
   include ExtractsHeadings
   include Linkable
-  include NationalApplicability
-  include Political
   include Updatable
+  include NationalApplicability
   include ActionView::Helpers::UrlHelper
 
   def body
     content_item["details"]["body"]
+  end
+
+  def breadcrumbs
+    return [] unless parent
+
+    e = parent
+    res = []
+
+    while e
+      res << { title: e["title"], url: e["base_path"] }
+      e = e["parent"] && e["parent"].first
+    end
+
+    res << { title: "Home", url: "/" }
+    res.reverse
   end
 
   def contents
@@ -27,5 +41,13 @@ class DetailedGuidePresenter < ContentItemPresenter
 
   def related_mainstream
     links("related_mainstream")
+  end
+
+private
+
+  def parent
+    if content_item["links"].include?("parent")
+      content_item["links"]["parent"][0]
+    end
   end
 end
