+++ app/presenters/html_publication_presenter.rb
@@ -38,4 +38,10 @@ class HtmlPublicationPresenter < ContentItemPresenter
     brand = "executive-office" if organisation["details"]["logo"]["crest"] == "eo"
     brand
   end
+
+private
+
+  def parent
+    content_item["links"]["parent"][0]
+  end
 end
