+++ app/presenters/topical_event_about_page_presenter.rb
@@ -1,5 +1,4 @@
 class TopicalEventAboutPagePresenter < ContentItemPresenter
-  include Breadcrumbs
   include ExtractsHeadings
   include ActionView::Helpers::UrlHelper
 
@@ -13,15 +12,27 @@ class TopicalEventAboutPagePresenter < ContentItemPresenter
     end
   end
 
+  def breadcrumbs
+    parent = topical_event
+    title = archived_topical_event? ? "#{parent['title']} (Archived)" : parent["title"]
+
+    [
+      { title: "Home", url: "/" },
+      { title: title, url: parent["base_path"] }
+    ]
+  end
+
 private
 
-  def parent
-    topical_event_end_date = super["details"]["end_date"]
+  def topical_event
+    content_item["links"]["parent"][0]
+  end
 
-    if topical_event_end_date && DateTime.parse(topical_event_end_date) <= Date.today
-      super.merge("title" => "#{super['title']} (Archived)")
-    else
-      super
-    end
+  def topical_event_end_date
+    topical_event["details"]["end_date"]
+  end
+
+  def archived_topical_event?
+    topical_event_end_date && DateTime.parse(topical_event_end_date) <= Date.today
   end
 end
