+++ app/views/content_items/case_study.html.erb
@@ -1,8 +1,6 @@
 <%= content_for :page_class, @content_item.format.dasherize.pluralize %>
 <%= content_for :title, "#{@content_item.page_title} - #{t("content_item.format.#{@content_item.document_type}", count: 1)}" %>
 
-<%= render "shared/breadcrumbs" %>
-
 <div class="grid-row">
   <div class="column-two-thirds">
     <%= render 'govuk_component/title',
