+++ app/views/content_items/detailed_guide.html.erb
@@ -1,7 +1,9 @@
 <%= content_for :page_class, @content_item.format.dasherize %>
 <%= content_for :title, @content_item.page_title %>
 
-<%= render "shared/breadcrumbs" %>
+<% if @content_item.breadcrumbs.any? %>
+  <%= render 'govuk_component/breadcrumbs', breadcrumbs: @content_item.breadcrumbs %>
+<% end %>
 
 <div class="grid-row">
   <div class="column-two-thirds">
