+++ app/views/content_items/document_collection.html.erb
@@ -1,8 +1,6 @@
 <%= content_for :page_class, @content_item.format.dasherize %>
 <%= content_for :title, @content_item.page_title %>
 
-<%= render "shared/breadcrumbs" %>
-
 <div class="grid-row">
   <div class="column-two-thirds">
     <%= render 'govuk_component/title',
