+++ app/views/content_items/topical_event_about_page.html.erb
@@ -1,7 +1,7 @@
 <%= content_for :page_class, @content_item.format.dasherize %>
 <%= content_for :title, @content_item.title %>
 
-<%= render "shared/breadcrumbs" %>
+<%= render 'govuk_component/breadcrumbs', breadcrumbs: @content_item.breadcrumbs %>
 
 <div class="grid-row">
   <div class="column-two-thirds">
