--- app/views/shared/_breadcrumbs.html.erb
@@ -1,3 +1,0 @@
-<% if @content_item.breadcrumbs.any? %>
-  <%= render 'govuk_component/breadcrumbs', breadcrumbs: @content_item.breadcrumbs %>
-<% end %>
